Rename LogoutButton to LogoutButtonRename
